Move ImageCarousel to component folder
