+++ src/templates/project-listing.js
@@ -3,7 +3,7 @@ import Layout from "../components/layout"
 import {getTitle} from "../utils/general"
 import Post from "../classes/post"
 import style from "./project-listing.module.css"
-import ImageCarousel from "../components/image-carousel"
+import ImageCarousel from "../components/ImageCarousel/ImageCarousel"
 import BigButton from "../components/big-button"
 import PostNavigation from "../components/post-navigation"
 import {graphql} from "gatsby"
